python/app-dynamic-sessions/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.5
    # via langchain
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==23.2.0
    # via aiohttp
azure-core==1.30.2
    # via azure-identity
azure-identity==1.17.0
    # via
    #   -r requirements.in
    #   langchain-azure-dynamic-sessions
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==42.0.8
    # via
    #   azure-identity
    #   msal
    #   pyjwt
distro==1.9.0
    # via openai
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.2.5
    # via -r requirements.in
langchain-azure-dynamic-sessions==0.1.0
    # via -r requirements.in
langchain-core==0.2.9
    # via
    #   langchain
    #   langchain-azure-dynamic-sessions
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.1.8
    # via -r requirements.in
langchain-text-splitters==0.2.1
    # via langchain
langchainhub==0.1.20
    # via -r requirements.in
langsmith==0.1.79
    # via
    #   langchain
    #   langchain-core
msal==1.28.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.1.0
    # via azure-identity
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
numpy==1.26.4
    # via langchain
openai==1.34.0
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.10.5
    # via langsmith
packaging==24.1
    # via
    #   langchain-core
    #   langchainhub
    #   msal-extensions
portalocker==2.8.2
    # via msal-extensions
pycparser==2.22
    # via cffi
pydantic==2.7.4
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.18.4
    # via pydantic
pyjwt[crypto]==2.8.0
    # via
    #   msal
    #   pyjwt
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-core
regex==2024.5.15
    # via tiktoken
requests==2.32.3
    # via
    #   azure-core
    #   langchain
    #   langchain-azure-dynamic-sessions
    #   langchainhub
    #   langsmith
    #   msal
    #   tiktoken
six==1.16.0
    # via azure-core
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.30
    # via langchain
tenacity==8.4.1
    # via
    #   langchain
    #   langchain-core
tiktoken==0.7.0
    # via langchain-openai
tqdm==4.66.4
    # via openai
types-requests==2.32.0.20240602
    # via langchainhub
typing-extensions==4.12.2
    # via
    #   azure-core
    #   azure-identity
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.2.2
    # via
    #   requests
    #   types-requests
yarl==1.9.4
    # via aiohttp
